+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add SAEec06 and SAEec09 SLOTs.
 - Add data page and extended data page to builder.
 - Implement `PartialEq` to mask of priority bits when comparing ids.
+- Fix spatial pointer flag not being set in memory request.
 
 ### Added